requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard
youtube_dl==2021.12.17
bing_image_downloader 
beautifulsoup4
bs4
cloudscraper
deepai
daxxhub
dnspython
emojis
ffmpeg-python
Faker
future==1.0.0
git+https://github.com/AsmSafone/SafoneAPI
gitpython
gpytranslate
gtts
hachoir
heroku3
httpx==0.27.0
img2pdf
instaloader
lexica-api
motor 
MukeshAPI
numpy
openai==0.28.0
opencv-python
pillow==10.3.0
psutil
pickledb
pornhub-api==0.3.0
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
pyrogram==2.0.106
python-dotenv
python-telegram-bot==21.1.1
python-whois
pyshorteners
pyttsx3
pytz
pytube
py-tgcalls==0.9.7
pyyaml==6.0.1
qrcode
requests
SafoneAPI
search_engine_parser
speedtest-cli
SpeechRecognition
spotipy
telegraph
tgcrypto
unidecode
wget
yt-dlp==2024.7.9
youtube-search
youtube-search-python
